+++ resources/attributes/default.rb
@@ -109,12 +109,12 @@ default["redborder"]["zookeeper_hosts"] = []
 default["redborder"]["memory_assigned"] = {}
 
 # geoip has been removed because is not a service
-default["redborder"]["services_group"]["full"] = %w[
-  consul chef-server zookeeper memcached rsyslog kafka logstash s3
-  redborder-ale postgresql redborder-postgresql nginx webui druid-broker
-  druid-historical druid-realtime druid-coordinator f2k redborder-monitor
-  pmacct redborder-dswatcher redborder-events-counter http2k
-  ]
+default["redborder"]["services_group"]["full"] = %w[consul chef-server zookeeper memcached rsyslog kafka logstash s3
+                                                    postgresql redborder-postgresql nginx webui druid-broker
+                                                    druid-historical druid-realtime druid-coordinator f2k
+                                                    redborder-monitor pmacct redborder-dswatcher
+                                                    redborder-events-counter http2k
+                                                    ]
 
 default["redborder"]["services_group"]["custom"] = []
 default["redborder"]["services_group"]["core"] = %w[consul zookeeper druid-coordinator druid-overlord hadoop-resourcemanager] #consul server
@@ -173,6 +173,7 @@ default["redborder"]["services"]["redborder-nmsp"]            = false
 default["redborder"]["services"]["redborder-ale"]             = false
 default["redborder"]["services"]["n2klocd"]                   = false
 default["redborder"]["services"]["radiusd"]                   = false
+default["redborder"]["services"]["postfix"]                   = true
 
 default["redborder"]["systemdservices"]["chef-client"]            = ["chef-client"]
 default["redborder"]["systemdservices"]["chef-server"]            = ["opscode-erchef"]
@@ -209,6 +210,7 @@ default["redborder"]["systemdservices"]["redborder-nmsp"]         = ["redborder-
 default["redborder"]["systemdservices"]["redborder-ale"]          = ["redborder-ale"]
 default["redborder"]["systemdservices"]["n2klocd"]                = ["n2klocd"]
 default["redborder"]["systemdservices"]["radiusd"]                = ["radiusd"]
+default["redborder"]["systemdservices"]["postfix"]                = ["postfix"]
 
 # Tier
 default["redborder"]["druid"]["historical"]["tier"]     = "default"